split ui.lua file into more specific files for each plugin
