+++ nvim/lua/bocha13/plugins/cmp.lua
@@ -1,0 +1,55 @@
+return {
+	{
+		"hrsh7th/nvim-cmp",
+		config = function()
+			local cmp = require("cmp")
+			local luasnip = require("luasnip")
+			local cmp_select = { behavior = cmp.SelectBehavior.Select }
+
+			cmp.setup({
+				snippet = {
+					expand = function(args)
+						luasnip.lsp_expand(args.body)
+					end,
+				},
+				sources = {
+					{ name = "copilot", group_index = 2 },
+					{ name = "luasnip", group_index = 3 },
+					{ name = "buffer", group_index = 2 },
+					{ name = "nvim_lsp", group_index = 2 },
+					{ name = "path", group_index = 2 },
+				},
+				mapping = cmp.mapping.preset.insert({
+					["<C-p>"] = cmp.mapping.select_prev_item(cmp_select),
+					["<C-n>"] = cmp.mapping.select_next_item(cmp_select),
+					["<C-y>"] = cmp.mapping.confirm({ select = true }),
+					["<C-Space>"] = cmp.mapping.complete(),
+					["<CR>"] = cmp.mapping.confirm({ select = false }),
+					["<Tab>"] = cmp.mapping(function(fallback)
+						if cmp.visible() then
+							cmp.select_next_item()
+						else
+							fallback()
+						end
+					end, {
+						"i",
+						"s",
+					}),
+				}),
+				window = {
+					documentation = {
+						winhighlight = "NormalFloat:Pmenu,NormalFloat:Pmenu,CursorLine:PmenuSel,Search:None",
+					},
+					completion = {
+						winhighlight = "NormalFloat:Pmenu,NormalFloat:Pmenu,CursorLine:PmenuSel,Search:None",
+					},
+				},
+			})
+		end,
+	},
+	{ "hrsh7th/cmp-nvim-lsp" },
+	{ "hrsh7th/cmp-buffer" },
+	{ "hrsh7th/cmp-path" },
+	{ "saadparwaiz1/cmp_luasnip" },
+	{ "hrsh7th/cmp-nvim-lua" },
+}
