+++ nvim/lua/bocha13/plugins/null-ls.lua
@@ -1,0 +1,41 @@
+return {
+	"jose-elias-alvarez/null-ls.nvim",
+	config = function()
+		local null_ls = require("null-ls")
+		local formatting = null_ls.builtins.formatting
+		local diagnostics = null_ls.builtins.diagnostics
+		local augroup = vim.api.nvim_create_augroup("LspFormatting", {})
+
+		null_ls.setup({
+			sources = {
+				formatting.prettier,
+				formatting.stylua,
+				diagnostics.eslint_d.with({
+					-- only enable islint if root has .eslintrc.js
+					confition = function(utils)
+						return utils.root_has_file(".eslintrc.js")
+					end,
+				}),
+			},
+			-- configure format on save
+			on_attach = function(current_client, bufnr)
+				if current_client.supports_method("textDocument/formatting") then
+					vim.api.nvim_clear_autocmds({ group = augroup, buffer = bufnr })
+					vim.api.nvim_create_autocmd("BufWritePre", {
+						group = augroup,
+						buffer = bufnr,
+						callback = function()
+							vim.lsp.buf.format({
+								filter = function(client)
+									--  only use null-ls for formatting instead of lsp server
+									return client.name == "null-ls"
+								end,
+								bufnr = bufnr,
+							})
+						end,
+					})
+				end
+			end,
+		})
+	end,
+}
